move selected message view state to separate package from message holder
